Add files via upload
Consider a snare drum with a diameter of 30 cm that is struck exactly in the centre, causing the centre of the drum to be
displaced downwards by 3 mm. Develop a model to describe the vibration of the drum skin, and then use Microsoft Excel to
create a numerical solution. Include a series of figures to depict the movement of the waves through time.